+++ NEWS.md
@@ -1,7 +1,7 @@
 # spduration 0.13.2
 
-* Added Belkin & Schofer 2003 coup data. 
-* Support for loglog in plot_hazard1.
+* Added Belkin & Schofer 2003 coup data in `data(bscoup)`. 
+* Replaced `plot_hazard1` and `plot_hazard2` with `plot_hazard`, added support for loglog models, and other internal streamlining. 
 * Partial fix for summarizing and printing results for models without intercept term.
 * Removed documentation for several internal functions.